Transition from std::array to vector and span view.
The reason to move from std::array is so that the sector class can take
responsibility for the file sectors that encode up a complete chunk.
This simplifies the design a great deal.  But it also requires that
the byte data be stored in a variably sized container -- or else a
variable number of fixed-size containers, which adds unnecessary
complications.